dispatcher: bit of refactoring (split PriorityURL)
